recruitment_agency/requirements.txt
agency-swarm>=0.4.4
python-dotenv>=0.19.0
requests>=2.26.0
pydantic>=2.0.0 